--- src/main/java/blog/Application.java
@@ -1,90 +1,0 @@
-package blog;
-
-import javax.sql.DataSource;
-
-import org.apache.ibatis.session.SqlSessionFactory;
-import org.apache.log4j.Logger;
-import org.mybatis.spring.SqlSessionFactoryBean;
-import org.mybatis.spring.annotation.MapperScan;
-import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.boot.SpringApplication;
-import org.springframework.boot.autoconfigure.EnableAutoConfiguration;
-import org.springframework.boot.autoconfigure.SpringBootApplication;
-import org.springframework.boot.autoconfigure.jdbc.DataSourceBuilder;
-import org.springframework.boot.context.embedded.ConfigurableEmbeddedServletContainer;
-import org.springframework.boot.context.embedded.EmbeddedServletContainerCustomizer;
-import org.springframework.boot.context.embedded.ErrorPage;
-import org.springframework.boot.context.properties.ConfigurationProperties;
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.ComponentScan;
-import org.springframework.context.annotation.Primary;
-import org.springframework.core.io.support.PathMatchingResourcePatternResolver;
-import org.springframework.http.HttpStatus;
-import org.springframework.jdbc.datasource.DataSourceTransactionManager;
-import org.springframework.transaction.PlatformTransactionManager;
-
-import com.mchange.v2.c3p0.ComboPooledDataSource;
-
-@EnableAutoConfiguration
-@SpringBootApplication
-@ComponentScan
-@MapperScan("blog.dao")
-// 该 @SpringBootApplication
-// 注解等价于以默认属性使用 @Configuration, @EnableAutoConfiguration和 @ComponentScan
-public class Application {
-    private static Logger logger = Logger.getLogger(Application.class);
-
-    // DataSource配置(C3P0)
-    @Bean(name = "dataSource")
-    @Qualifier(value = "dataSource")
-    @Primary
-    @ConfigurationProperties(prefix = "c3p0")
-    public DataSource dataSource() {
-        return DataSourceBuilder.create().type(ComboPooledDataSource.class)
-                .build();
-    }
-
-    // 提供SqlSeesion
-    @Bean
-    public SqlSessionFactory sqlSessionFactoryBean() throws Exception {
-        SqlSessionFactoryBean sqlSessionFactoryBean = new SqlSessionFactoryBean();
-        sqlSessionFactoryBean.setDataSource(dataSource());
-        PathMatchingResourcePatternResolver resolver = new PathMatchingResourcePatternResolver();
-        sqlSessionFactoryBean.setMapperLocations(resolver.getResources("classpath:/MyBatis_Mapper/*.xml"));
-        return sqlSessionFactoryBean.getObject();
-    }
-
-    @Bean   //事务???不太清楚,应该是
-    public PlatformTransactionManager transactionManager() {
-        return new DataSourceTransactionManager(dataSource());
-    }
-    
-    /**
-     * 错误页面的处理
-     * @return
-     */
-    @Bean
-    public EmbeddedServletContainerCustomizer containerCustomizer() {
-
-        return new EmbeddedServletContainerCustomizer() {
-            @Override
-            public void customize(ConfigurableEmbeddedServletContainer container) {
-                ErrorPage error401Page = new ErrorPage(HttpStatus.UNAUTHORIZED, "/401.html");
-                ErrorPage error404Page = new ErrorPage(HttpStatus.NOT_FOUND, "/404.html");
-                ErrorPage error500Page = new ErrorPage(HttpStatus.INTERNAL_SERVER_ERROR, "/500.html");
-                container.addErrorPages(error401Page, error404Page, error500Page);
-            }
-        };
-    }
-    
-    
-
-    /**
-     * Main Start
-     */
-    public static void main(String[] args) {
-        SpringApplication.run(Application.class, args);
-        logger.info("============= SpringBoot Start Success =============");
-    }
-
-}
