--- src/main/java/blog/dao/IStudentDao.java
@@ -1,10 +1,0 @@
-package blog.dao;
-
-import java.util.List;
-
-import blog.pojo.Student;
-
-public interface IStudentDao {
-    public List<Student> findStudenInfo();
-
-}
