--- src/main/java/blog/service/IStudentService.java
@@ -1,9 +1,0 @@
-package blog.service;
-
-import blog.pojo.Student;
-
-public interface IStudentService {
-
-    public Student getUserInfo();
-
-}
