+++ src/main/java/pet/dao/IStudentDao.java
@@ -1,0 +1,11 @@
+package pet.dao;
+
+import java.util.List;
+
+import pet.model.Student;
+
+
+public interface IStudentDao {
+    public List<Student> findStudenInfo();
+
+}
