+++ src/main/java/pet/service/IStudentService.java
@@ -1,0 +1,9 @@
+package pet.service;
+
+import pet.model.Student;
+
+public interface IStudentService {
+
+    public Student getUserInfo();
+
+}
